replace BigDecimal with Double (#815)
* replace BigDecimal with Double

* remove unused import

+++ src/main/scala/beam/agentsim/agents/choice/mode/ModeChoiceTransitIfAvailable.scala
@@ -28,12 +28,7 @@ class ModeChoiceTransitIfAvailable(val beamServices: BeamServices) extends ModeC
 
   override def utilityOf(alternative: EmbodiedBeamTrip): Double = 0.0
 
-  override def utilityOf(
-    mode: Modes.BeamMode,
-    cost: BigDecimal,
-    time: BigDecimal,
-    numTransfers: Int
-  ): Double = 0.0
+  override def utilityOf(mode: Modes.BeamMode, cost: Double, time: Double, numTransfers: Int): Double = 0.0
 }
 
 object ModeChoiceTransitIfAvailable {}
